+++ packages/core/src/manglers/css-classes.ts
@@ -182,7 +182,11 @@ export default class CssClassMangler extends SimpleManglerPlugin {
   private static getPatterns(
     classNamePattern?: string | string[],
   ): string | string[] {
-    return classNamePattern || CssClassMangler.DEFAULT_PATTERNS;
+    if (classNamePattern === undefined) {
+      return CssClassMangler.DEFAULT_PATTERNS;
+    }
+
+    return classNamePattern;
   }
 
   /**
@@ -192,7 +196,11 @@ export default class CssClassMangler extends SimpleManglerPlugin {
    * @returns The reserved names to be used.
    */
   private static getReserved(reservedClassNames?: string[]): string[] {
-    return reservedClassNames || CssClassMangler.DEFAULT_RESERVED;
+    if (reservedClassNames === undefined) {
+      return CssClassMangler.DEFAULT_RESERVED;
+    }
+
+    return reservedClassNames;
   }
 
   /**
@@ -202,6 +210,10 @@ export default class CssClassMangler extends SimpleManglerPlugin {
    * @returns The prefix to be used.
    */
   private static getPrefix(keepClassNamePrefix?: string): string {
-    return keepClassNamePrefix || CssClassMangler.DEFAULT_PREFIX;
+    if (keepClassNamePrefix === undefined) {
+      return CssClassMangler.DEFAULT_PREFIX;
+    }
+
+    return keepClassNamePrefix;
   }
 }