+++ packages/core/src/manglers/css-variables.ts
@@ -180,7 +180,11 @@ export default class CssVariableMangler extends SimpleManglerPlugin {
   private static getPatterns(
     cssVarNamePattern?: string | string[],
   ): string | string[] {
-    return cssVarNamePattern || CssVariableMangler.DEFAULT_PATTERNS;
+    if (cssVarNamePattern === undefined) {
+      return CssVariableMangler.DEFAULT_PATTERNS;
+    }
+
+    return cssVarNamePattern;
   }
 
   /**
@@ -190,7 +194,11 @@ export default class CssVariableMangler extends SimpleManglerPlugin {
    * @returns The reserved names to be used.
    */
   private static getReserved(reservedCssVarNames?: string[]): string[] {
-    return reservedCssVarNames || CssVariableMangler.DEFAULT_RESERVED;
+    if (reservedCssVarNames === undefined) {
+      return CssVariableMangler.DEFAULT_RESERVED;
+    }
+
+    return reservedCssVarNames;
   }
 
   /**
@@ -200,6 +208,10 @@ export default class CssVariableMangler extends SimpleManglerPlugin {
    * @returns The prefix to be used.
    */
   private static getPrefix(keepCssVarPrefix?: string): string {
-    return keepCssVarPrefix || CssVariableMangler.DEFAULT_PREFIX;
+    if (keepCssVarPrefix === undefined) {
+      return CssVariableMangler.DEFAULT_PREFIX;
+    }
+
+    return keepCssVarPrefix;
   }
 }