+++ packages/core/src/manglers/html-attributes.ts
@@ -196,7 +196,11 @@ export default class HtmlAttributeMangler extends SimpleManglerPlugin {
   private static getPatterns(
     attrNamePattern?: string | string[],
   ): string | string[] {
-    return attrNamePattern || HtmlAttributeMangler.DEFAULT_PATTERNS;
+    if (attrNamePattern === undefined) {
+      return HtmlAttributeMangler.DEFAULT_PATTERNS;
+    }
+
+    return attrNamePattern;
   }
 
   /**
@@ -206,7 +210,11 @@ export default class HtmlAttributeMangler extends SimpleManglerPlugin {
    * @returns The reserved names to be used.
    */
   private static getReserved(reservedAttrNames?: string[]): string[] {
-    return reservedAttrNames || HtmlAttributeMangler.DEFAULT_RESERVED;
+    if (reservedAttrNames === undefined) {
+      return HtmlAttributeMangler.DEFAULT_RESERVED;
+    }
+
+    return reservedAttrNames;
   }
 
   /**
@@ -216,6 +224,10 @@ export default class HtmlAttributeMangler extends SimpleManglerPlugin {
    * @returns The prefix to be used.
    */
   private static getPrefix(keepAttrPrefix?: string): string {
-    return keepAttrPrefix || HtmlAttributeMangler.DEFAULT_PREFIX;
+    if (keepAttrPrefix === undefined) {
+      return HtmlAttributeMangler.DEFAULT_PREFIX;
+    }
+
+    return keepAttrPrefix;
   }
 }